pyproject.toml
[project]
name = "beyond-chatgpt"
version = "0.1.0"
description = "gpt-like-a-dev"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.12",
    "jupyter>=1.1.1",
    "openai",
    "pydantic>=2.11.4",
    "uvicorn>=0.34.2",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-redis>=0.3.0",
    "qdrant-client>=1.12.0",
    "pypdf>=5.0.0",
]
